feat: mov all types to the Interface.tsx file
