+++ src/components/ErrorButton/ErrorButton.tsx
@@ -1,9 +1,6 @@
 import { FC } from 'react';
+import { ErrorButtonProps } from '../../types/Interface';
 import styles from './ErrorButton.module.scss';
-
-interface ErrorButtonProps {
-  onClick: () => void;
-}
 
 const ErrorButton: FC<ErrorButtonProps> = ({ onClick }) => {
   return (
