+++ src/components/Header/Header.module.scss
@@ -39,8 +39,6 @@
   left: 0;
   right: 0;
   bottom: 0;
-  background: rgba(0, 0, 0, 0.5);
-  z-index: 10;
 }
 
 @media (max-width: 768px) {
