+++ src/components/Layout/Layout.tsx
@@ -3,12 +3,9 @@ import { Outlet } from 'react-router-dom';
 import Footer from '../Footer/Footer';
 import Header from '../Header/Header';
 import style from './Layout.module.scss';
+import { LayoutProps } from '../../types/Interface';
 
-interface Props {
-  className?: string;
-}
-
-export const Layout: React.FC<Props> = () => {
+export const Layout: React.FC<LayoutProps> = () => {
   return (
     <>
       <Header />
